@@ -1,17 +1,18 @@
 {
   "name": "@leadoo/react-bot",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "description": "React components for Leadoo bots",
-  "main": "dist/index.js",
+  "type": "module",
+  "exports": "./dist/index.js",
   "typings": "dist/index.d.ts",
   "scripts": {
     "build": "run-s clean build:web",
-    "build:web": "webpack --mode production --config webpack.config.js",
+    "build:web": "webpack --mode production --config webpack.config.cjs",
     "clean": "rimraf ./dist",
     "dev": "npm run clean && webpack --mode development --config webpack.config.js --watch",
     "format": "prettier --write '{source,stories,test}/**/*.{ts,tsx,js,jsx}'",
     "prepublishOnly": "npm run build",
-    "storybook": "NODE_PATH=source start-storybook",
+    "storybook": "storybook dev -p 6008 -s public",
     "test": "run-s test:specs test:format",
     "test:format": "prettier-check '{source,stories,test}/**/*.{ts,tsx,js,jsx}'",
     "test:specs": "jest"
@@ -20,6 +21,7 @@
     "dist/**/*",
     "*.md"
   ],
+  "sideEffects": false,
   "lint-staged": {
     "{source,stories,test}/**/*.{ts,tsx,js,jsx}": [
       "prettier --write"
@@ -35,7 +37,30 @@
     "url": "git+https://github.com/LeadooMT/react-bot.git"
   },
   "jest": {
-    "preset": "ts-jest"
+    "extensionsToTreatAsEsm": [
+      ".ts",
+      ".tsx",
+      ".jsx"
+    ],
+    "moduleNameMapper": {
+      "^(\\.{1,2}/.*)\\.jsx$": "$1",
+      "^(\\.{1,2}/.*)\\.js$": "$1"
+    },
+    "preset": "ts-jest/presets/default-esm",
+    "setupFilesAfterEnv": [
+      "<rootDir>/test/jest-setup.ts"
+    ],
+    "testEnvironment": "jsdom",
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest",
+      "^.+\\.tsx?$": [
+        "ts-jest",
+        {
+          "tsconfig": "tsconfig.json",
+          "useESM": true
+        }
+      ]
+    }
   },
   "keywords": [
     "leadoo",
@@ -50,32 +75,38 @@
   },
   "homepage": "https://github.com/LeadooMT/react-bot#readme",
   "devDependencies": {
-    "@babel/core": "^7.11.0",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.10.4",
-    "@storybook/react": "^5.3.19",
-    "@types/jest": "^26.0.9",
-    "@types/react": "^16.9.44",
-    "@types/react-dom": "^16.9.8",
-    "babel-jest": "^26.2.2",
-    "babel-loader": "^8.1.0",
-    "husky": "^4.2.5",
-    "jest": "^26.2.2",
-    "lint-staged": "^10.2.11",
+    "@babel/core": "^7.23.5",
+    "@babel/preset-env": "^7.23.5",
+    "@babel/preset-react": "^7.23.3",
+    "@storybook/react": "^7.6.2",
+    "@storybook/react-webpack5": "^7.6.2",
+    "@testing-library/jest-dom": "^6.1.5",
+    "@testing-library/react": "^14.1.2",
+    "@types/jest": "^29.5.10",
+    "@types/react": "^18.2.39",
+    "@types/react-dom": "^18.2.17",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.1.3",
+    "husky": "^4.3.8",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^15.1.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.19.1",
+    "prettier": "^3.1.0",
     "prettier-check": "^2.0.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "rimraf": "^3.0.2",
-    "ts-jest": "^26.1.4",
-    "ts-loader": "^8.0.2",
-    "typescript": "^3.9.7",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "resolve-typescript-plugin": "^2.0.1",
+    "rimraf": "^5.0.5",
+    "storybook": "^7.6.2",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.3.2",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
   },
   "peerDependencies": {
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
