@@ -1,5 +1,5 @@
 import * as React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { act } from "react-dom/test-utils";
 import { VisualBot } from "../source/VisualBot";
 
@@ -25,10 +25,8 @@ afterEach(() => {
 describe("VisualBot", () => {
     test("inserts script", async () => {
         act(() => {
-            ReactDOM.render(
-                <VisualBot code="test" generateScript={fakeScriptInserter} />,
-                container
-            );
+            const root = createRoot(container);
+            root.render(<VisualBot code="test" generateScript={fakeScriptInserter} />);
         });
         const divContainer = container.getElementsByTagName("div")[0];
         const script = divContainer.children[0];
@@ -37,10 +35,8 @@ describe("VisualBot", () => {
 
     test("sets correct standard URL", async () => {
         act(() => {
-            ReactDOM.render(
-                <VisualBot code="test" generateScript={fakeScriptInserter} />,
-                container
-            );
+            const root = createRoot(container);
+            root.render(<VisualBot code="test" generateScript={fakeScriptInserter} />);
         });
         const script = container.getElementsByTagName("script")[0];
         expect(script.dataset.url).toMatch(/bot\.leadoo\.com/);
@@ -49,10 +45,8 @@ describe("VisualBot", () => {
 
     test("includes seamless flag in URL", async () => {
         act(() => {
-            ReactDOM.render(
-                <VisualBot code="test" generateScript={fakeScriptInserter} seamless />,
-                container
-            );
+            const root = createRoot(container);
+            root.render(<VisualBot code="test" generateScript={fakeScriptInserter} seamless />);
         });
         const script = container.getElementsByTagName("script")[0];
         expect(script.dataset.url).toMatch(/bot\.leadoo\.com/);
