@@ -1,5 +1,5 @@
-import * as React from "react";
-import { InPageBot } from "../source/index";
+import React from "react";
+import { InPageBot } from "../source/index.js";
 
 const DEMO_MEDIA = "TestMediaCode";
 const DEMO_MEDIA_BOT_CODE = "react-bot";
