@@ -1,12 +1,19 @@
 const path = require("path");
+const ResolveTypeScriptPlugin =  require("resolve-typescript-plugin");
 const pkg = require('./package.json');
 
 module.exports = {
-    entry: path.resolve(__dirname, "./source/index.ts"),
+    entry: path.resolve(__dirname, "./source/index.js"),
+
+    experiments: {
+        outputModule: true,
+    },
 
     externals: [
         ...Object.keys(pkg.peerDependencies)
     ],
+
+    externalsType: "module",
 
     module: {
         rules: [
@@ -20,9 +27,7 @@ module.exports = {
                             presets: [
                                 "@babel/preset-react",
                                 ["@babel/preset-env", {
-                                    "targets": {
-                                        "ie": 11
-                                    },
+                                    "targets": "> 0.5%, not dead",
                                     "useBuiltIns": false
                                 }]
                             ]
@@ -31,7 +36,10 @@ module.exports = {
                     {
                         loader: "ts-loader"
                     }
-                ]
+                ],
+                resolve: {
+                    fullySpecified: false
+                }
             }
         ]
     },
@@ -39,10 +47,21 @@ module.exports = {
     output: {
         filename: "index.js",
         path: path.resolve(__dirname, "./dist"),
-        libraryTarget: "commonjs2"
+        environment: {
+            module: true
+        },
+        library: {
+            type: "module"
+        }
     },
 
     resolve: {
-        extensions: [".ts", ".tsx"]
-    }
+        extensions: [".js", ".jsx"],
+        plugins: [
+            // Handle .ts => .js resolution
+            new ResolveTypeScriptPlugin()
+        ]
+    },
+
+    target: "web"
 };
